Receive user choices and update game board
Using a loop, alternates the play between each player one move at a
time. It will ask the user for an input (#1-9), and if the spot they
choose is not empty, it will update the game board with their piece
(whether they are X or O). Then the screen will clear and it will be the
next users turn.

The game currently loops forever and cannot end. The win condition
is not coded yet.